Chore: Move Home page to pages
auth 내부의 Home page pages 내부로 이동
